profile creation and profile edit logic with API endpoints
