+++ src/client/pages/Home.jsx
@@ -1,5 +1,4 @@
 
-
 export default function Home(){
     return (
         <main>
